Reverted back GetSentenceWidth to original code and made it so 0x0A is now 0x0A0D in the menus and set the width for them to 2.

+++ code/iblard/code/font.h
@@ -9,8 +9,7 @@ extern "C" {
 
 	u32 GetLetterWidth(u32 letter);
 
-	u32 GetSentenceWidth(const u8* text, const u8* end);
-
+	u32 GetSentenceWidth(const u16* text, const u16* end);
 }
 
 #endif
